requirements.txt
streamlit
langchain
langchain-ollama
langchain-pinecone
pinecone-client
langchain-google-vertexai
tqdm
python-dotenv
beautifulsoup4
requests
urllib3